Improve test file structure
